migrated some functions to tokenization.util
- fixed a bunch of imports
- modified `get_tokens_up_to_path_start` to depend on tokenization mode
