+++ maze_dataset/tokenization/__init__.py
@@ -1,6 +1,6 @@
 from maze_dataset.tokenization.maze_tokenizer import MazeTokenizer, TokenizationMode
+from maze_dataset.tokenization.util import coord_str_to_tuple
 from maze_dataset.tokenization.token_utils import (
-    coord_str_to_tuple,
     get_adj_list_tokens,
     get_context_tokens,
     get_origin_tokens,
