+++ tests/unit/maze_dataset/generation/test_coord_str_tuple.py
@@ -1,7 +1,7 @@
 import numpy as np
 import pytest
 
-from maze_dataset.tokenization.token_utils import (
+from maze_dataset.tokenization.util import (
     _coord_to_strings_indexed,
     _coord_to_strings_UT,
     coord_str_to_coord_np,
