+++ tests/unit/maze_dataset/tokenization/test_token_utils.py
@@ -2,15 +2,17 @@ import pytest
 from pytest import mark, param
 
 from maze_dataset.dataset.maze_dataset import MazeDatasetConfig
-from maze_dataset.tokenization.token_utils import (
+from maze_dataset.tokenization.util import (
     _coord_to_strings_UT,
     coords_to_strings,
+    strings_to_coords,
+)
+from maze_dataset.tokenization.token_utils import (
     get_adj_list_tokens,
     get_origin_tokens,
     get_path_tokens,
     get_target_tokens,
     get_tokens_up_to_path_start,
-    strings_to_coords,
     tokens_between,
 )
